refactor: migrate server from JS to TS
