+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id?: string;
+  username: string;
+  age: number;
+  hobbies: string[];
+}
