--- src/users.js
@@ -1,3 +1,0 @@
-const users = {};
-
-export default users;
