+++ src/users.ts
@@ -1,0 +1,5 @@
+import { User } from "./types/index";
+
+const users: User[] = [];
+
+export default users;
