+++ src/utils/is-valid-uuid.ts
@@ -1,4 +1,4 @@
-const isValidUuid = (uuid) => {
+const isValidUuid = (uuid: string) => {
   const uuidRE = new RegExp(
     /^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$/i
   );
